「OpenHarmony贡献者计划2022」update engine/scene/recorder/standard/hirecorder_impl.cpp. link #I5C5I0 Signed-off-by: mr-yx <496043997@qq.com>

+++ engine/scene/recorder/standard/hirecorder_impl.cpp
@@ -92,7 +92,7 @@ int32_t HiRecorderImpl::SetAudioSource(AudioSourceType source, int32_t& sourceId
     if (ret == ErrorCode::SUCCESS) {
         audioCount_++;
         audioSourceId_ = tempSourceId;
-        sourceId = audioSourceId_;
+        sourceId = static_cast<int32_t>(audioSourceId_);
     }
     PROFILE_END("SetAudioSource end.");
     return TransErrorCode(ret);
